% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{average_lifespan_occurrence}
\alias{average_lifespan_occurrence}
\title{average_lifespan_occurrence}
\usage{
average_lifespan_occurrence(x, y)
}
\arguments{
\item{y}{}
}
\description{
average_lifespan_occurrence
}
